requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Brotli==1.0.9
cattrs==1.10.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
docopt==0.6.2
executing==0.8.3
ffmpeg==1.4
frozenlist==1.3.0
idna==3.3
iso8601==1.0.2
Js2Py==0.71
multidict==6.0.2
mutagen==1.45.1
packaging==21.3
pickleshare==0.7.5
pipwin==0.5.2
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
pycryptodomex==3.14.1
Pygments==2.12.0
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyzmq==23.1.0
requests==2.28.0
requests-cache==0.9.4
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
traitlets==5.2.2.post1
twitchio==2.3.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
url-normalize==1.4.3
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.3
winrt==1.0.21033.1
yarl==1.7.2
yt-dlp==2022.5.18
